Rename actions to croquemort, refs #143
